+++ test/thread_safety.jl
@@ -39,8 +39,7 @@ let
 
     for agg in (VR_FRM(), VR_Direct(), VR_DirectFW())
         jump_prob = JumpProblem(ode_prob, Direct(), vrj; vr_aggregator = agg)
-        prob_func(prob, i, repeat) = deepcopy(prob)
-        prob = EnsembleProblem(jump_prob, prob_func = prob_func)
+        prob = EnsembleProblem(jump_prob)
         sol = solve(prob, Tsit5(), EnsembleThreads(), trajectories = 400,
             save_everystep = false)
         firstrx_time = [sol.u[i].t[findfirst(>(sol.u[i].t[1]), sol.u[i].t)] for i in 1:length(sol)]
@@ -58,8 +57,7 @@ let
 
     for agg in (VR_FRM(), VR_Direct(), VR_DirectFW())
         jump_prob = JumpProblem(sde_prob, Direct(), vrj; vr_aggregator = agg)
-        prob_func(prob, i, repeat) = deepcopy(prob)
-        prob = EnsembleProblem(jump_prob; prob_func)
+        prob = EnsembleProblem(jump_prob)
         sol = solve(prob, SRIW1(), EnsembleThreads();
             trajectories = 400, save_everystep = false)
         firstrx_time = [sol.u[i].t[findfirst(>(sol.u[i].t[1]), sol.u[i].t)] for i in 1:length(sol)]